add utility class
